+++ packages/stylelint-config-nikita/index.js
@@ -95,11 +95,12 @@ module.exports = {
         "media-query-list-comma-space-after": "always-single-line",
         "media-query-list-comma-space-before": "never",
         "no-empty-source": true,
-        "no-eol-whitespace": [true, { ignore: ["empty-lines"] }], //remove option if no-eol-whitespace is fixable
+        "no-eol-whitespace": null, //remove option if no-eol-whitespace is fixable
+        // "no-eol-whitespace": [false, { ignore: ["empty-lines"] }], //remove option if no-eol-whitespace is fixable
         "no-extra-semicolons": true,
         "no-invalid-double-slash-comments": true,
         "no-missing-end-of-source-newline": true,
-        "number-leading-zero": "always",
+        "number-leading-zero": "never",
         "number-no-trailing-zeros": true,
         "property-case": "lower",
         "property-no-unknown": true,